added user-specific content

+++ express-server.js
@@ -11,8 +11,14 @@ app.use(cookieParser());
 app.set("view engine", "ejs");
 
 var urlDatabase = {
-  "b2xVn2": "http://www.lighthouselabs.ca",
-  "9sm5xK": "http://www.google.com"
+  "b2xVn2": {
+    originalURL: "http://www.lighthouselabs.ca",
+    userID: "userRandomID"
+  },
+  "9sm5xK": {
+    originalURL: "http://www.google.com",
+    userID: "user2RandomID"
+  }
 };
 
 const users = {
@@ -43,16 +49,17 @@ app.get("/login", (req, res) => {
 });
 
 app.post("/login", (req, res) => {
+  console.log(users);
   if (req.body.email === "" || req.body.password === "") {
     res.sendStatus(400);
   }
   let user;
-  for (var key in users) {
+  for (let key in users) {
     if (users[key].email === req.body.email) {
     user = users[key].id;
     }
   }
-  if (user && users[key].password === req.body.password) {
+  if (user && users[user].password === req.body.password) {
     res.cookie("user_id", user);
     res.redirect("/urls");
   } else {
@@ -64,6 +71,8 @@ app.post("/logout", (req, res) => {
   res.clearCookie("user_id");
   res.redirect("/urls");
 });
+
+// urlsForUSER BROKESHIT
 
 app.post("/register", (req, res) => {
   if (req.body.email === "" || req.body.password === "") {
@@ -83,7 +92,11 @@ app.post("/register", (req, res) => {
 
 app.get("/urls/new", (req, res) => {
   let templateVars = { users: users[req.cookies.user_id] };
-  res.render("urls_new", templateVars);
+  if (!req.cookies.user_id) {
+    res.redirect("/register");
+  } else {
+    res.render("urls_new", templateVars);
+  }
 });
 
 app.get("/urls.json", (req, res) => {
@@ -91,13 +104,19 @@ app.get("/urls.json", (req, res) => {
 });
 
 app.get("/urls", (req, res) => {
-  let templateVars = { urls: urlDatabase, users: users[req.cookies.user_id] };
+  let templateVars = { urls: urlsForUser(req.cookies.user_id), users: users[req.cookies.user_id] };
+  console.log(templateVars)
   res.render("urls_index", templateVars);
 });
 
 app.get("/urls/:id", (req, res) => {
-  let templateVars = { shortURL: req.params.id, longURL: urlDatabase[req.params.id], users: users[req.cookies.user_id] };
-  res.render("urls_show", templateVars);
+  let templateVars = { shortURL: req.params.id, longURL: urlDatabase[req.params.id].originalURL, users: users[req.cookies.user_id] };
+  if (req.cookies.user_id !== urlDatabase[req.params.id].userID) {
+    res.sendStatus(403);
+  // res.render("urls_show", templateVars);
+  } else {
+    res.render("urls_show", templateVars);
+  }
 });
 
 app.get("/hello", (req, res) => {
@@ -105,26 +124,36 @@ app.get("/hello", (req, res) => {
 });
 
 app.get("/u/:shortURL", (req, res) => {
-  let longURL = urlDatabase[req.params.shortURL];
+  let longURL = urlDatabase[req.params.shortURL].originalURL;
   // console.log(longURL);
   res.redirect(longURL);
 });
 
 app.post("/urls", (req, res) => {
   var shortURLKey = generateRandomString();
-  urlDatabase[shortURLKey] = req.body.longURL;
+  console.log("New URL : " + req.body.longURL + " User ID: " + req.cookies.user_id);
+  urlDatabase[shortURLKey] = {originalURL: req.body.longURL, userID: req.cookies.user_id};
   res.redirect("/urls");
 });
 
 app.post("/urls/:id/delete", (req, res) => {
-  // console.log(req.params.id);
-  deleteURL(req.params.id);
-  res.redirect("/urls");
+  console.log(urlDatabase);
+  if (req.cookies.user_id === urlDatabase[req.params.id].userID) {
+    deleteURL(req.params.id);
+    res.redirect("/urls");
+  } else {
+    res.sendStatus(403);
+  }
 });
 
 app.post("/urls/:id/update", (req, res) => {
-  updateURL(req.params.id, req.body.longURL);
-  res.redirect("/urls");
+  if (req.cookies.user_id === urlDatabase[req.params.id].userID) {
+    updateURL(req.params.id, req.body.longURL, req.cookies.user_id);
+    res.redirect("/urls");
+  } else {
+    res.sendStatus(403);
+  }
+
 });
 
 function generateRandomString() {
@@ -135,14 +164,29 @@ function deleteURL(id) {
   delete urlDatabase[id];
 }
 
-function updateURL(id, newURL) {
+function updateURL(id, newURL, userID) {
   deleteURL(id);
-  urlDatabase[id] = newURL;
+  urlDatabase[id] = {originalURL: newURL, userID: userID};
+  // console.log(urlDatabase);
 }
 
 function createUser(id, email, password) {
   // console.log("Email: " + email + " Password: " + password);
   users[id] = {id: id, email: email, password: password};
+}
+
+function urlsForUser(id) {
+  var filteredURLs = {};
+  for (var key in urlDatabase) {
+    if (urlDatabase[key].userID === id) {
+      filteredURLs[key] = {
+        longURL: urlDatabase[key].originalURL,
+        shortURL: key,
+        userID: id
+      }
+    }
+  }
+return filteredURLs;
 }
 
 app.listen(PORT, () => {
